Move protocol validation to url utils
