+++ src/app/pages/NFTInstanceDashboardPage/InstanceImageCard.tsx
@@ -11,7 +11,8 @@ import Tooltip from '@mui/material/Tooltip'
 import OpenInFullIcon from '@mui/icons-material/OpenInFull'
 import { styled } from '@mui/material/styles'
 import { EvmNft } from 'oasis-nexus/api'
-import { isNftImageUrlValid, processNftImageUrl } from '../../utils/nft-images'
+import { processNftImageUrl } from '../../utils/nft-images'
+import { hasValidProtocol } from '../../utils/url'
 import { COLORS } from '../../../styles/theme/colors'
 import { ImagePreview } from '../../components/ImagePreview'
 import { NoPreview } from '../../components/NoPreview'
@@ -112,7 +113,7 @@ export const InstanceImageCard: FC<InstanceImageCardProps> = ({ isFetched, isLoa
                 }}
               >
                 <NoPreview placeholderSize={imageSize} />
-                {isNftImageUrlValid(nft.image) && (
+                {hasValidProtocol(nft.image) && (
                   <Link href={nft.image} rel="noopener noreferrer" target="_blank">
                     {t('nft.openInNewTab')}
                   </Link>
@@ -120,7 +121,7 @@ export const InstanceImageCard: FC<InstanceImageCardProps> = ({ isFetched, isLoa
               </Box>
             </Box>
           )}
-          {isFetched && nft && isNftImageUrlValid(nft.image) && !imageLoadError && (
+          {isFetched && nft && hasValidProtocol(nft.image) && !imageLoadError && (
             <>
               <Box
                 sx={{
