+++ src/app/pages/TokenDashboardPage/ImageListItemImage.tsx
@@ -5,7 +5,8 @@ import Box from '@mui/material/Box'
 import Link from '@mui/material/Link'
 import OpenInBrowserIcon from '@mui/icons-material/OpenInBrowser'
 import { styled } from '@mui/material/styles'
-import { isNftImageUrlValid, processNftImageUrl } from 'app/utils/nft-images'
+import { processNftImageUrl } from 'app/utils/nft-images'
+import { hasValidProtocol } from 'app/utils/url'
 import { NoPreview } from '../../components/NoPreview'
 import { getNftInstanceLabel } from '../../utils/nft'
 import { EvmNft } from '../../../oasis-nexus/api'
@@ -61,7 +62,7 @@ export const ImageListItemImage: FC<ImageListItemImageProps> = ({ instance, to }
 
   return (
     <Link component={RouterLink} to={to} sx={{ display: 'flex', position: 'relative' }}>
-      {isNftImageUrlValid(instance.image) && !imageLoadError ? (
+      {hasValidProtocol(instance.image) && !imageLoadError ? (
         <StyledImage
           onError={() => setImageLoadError(true)}
           src={processNftImageUrl(instance.image)}
