+++ src/app/utils/nft-images.ts
@@ -1,23 +1,9 @@
 import { AppError, AppErrors } from 'types/errors'
 import { accessIpfsUrl } from './ipfs'
-
-const validProtocols = ['http:', 'https:', 'ftp:', 'ipfs:', 'data:']
-
-export const isNftImageUrlValid = (url: string | undefined): boolean => {
-  if (!url) {
-    return false
-  }
-
-  try {
-    const parsedUrl = new URL(url)
-    return validProtocols.includes(parsedUrl.protocol)
-  } catch (error) {
-    return false
-  }
-}
+import { hasValidProtocol } from './url'
 
 export const processNftImageUrl = (url: string | undefined): string => {
-  if (!url || !isNftImageUrlValid(url)) {
+  if (!url || !hasValidProtocol(url)) {
     throw new AppError(AppErrors.InvalidUrl)
   }
